+++ src/SpritePreviewer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import './SpritePreviewer.css';
-import { customCharacters, getPercent, getUrl } from './sprites.js';
+import { customCharacters, customVanillas, getPercent, getUrl } from './sprites.js';
 
 const containerStyle = {
   display: "flex",
@@ -9,15 +9,17 @@ const containerStyle = {
   justifyContent: "space-around",
 } 
 
-const boxStyle = {
+const boxStyle = (mode) => { return {
     display: "flex",
     borderRadius: "4px",
     border: "1px solid #bbb",
     padding: "4px",
     margin: "2px",
-    flexBasis: "32%",
-    minWidth: "175px",
-}
+    flexBasis: mode === 1 ? "32%": "19%",
+    minWidth: mode === 1 ? "175px" : "75px",
+    flexDirection: mode === 1 ? "row" : "column-reverse",
+    alignItems: mode === 1 ? "inherit" : "center",
+}}
 
 const spriteDisplayStyle = (character) => { return {
     backgroundImage: `url('${getUrl(character?.value, 1)}')`,
@@ -48,17 +50,24 @@ const conversionStyle = {
     fontSize: "0.8em",
 }
 
-const SpritePreviewer = () => {
+
+const SpritePreviewer = (props) => {
+    const customs = props.mode === 1 ? customCharacters : customVanillas;
     return <div style={containerStyle}>
-    { customCharacters.map(character => {
-        return <div key={character.value} style={boxStyle}>
+    { customs === customVanillas && <div>
+        Expanded vanilla sprites have been improved, with alternate sprite versions—ghost, climbing, jumping, etc—by Defqon1 (thank you!)
+    </div>}
+    { customs.map(character => {
+        return <div key={character.value} style={boxStyle(props.mode)}>
             <div className="animatedSprite" style={spriteDisplayStyle(character)} />
             <div style={infoStyle}>
                 <span style={labelStyle}>{character.label}</span>
-                <span style={creatorStyle}>Creator: {character.creator}</span>
-                <span style={conversionStyle}>
-                    {getPercent(character)}% conversion
-                </span>
+                { customs === customCharacters && <>
+                    <span style={creatorStyle}>Creator: {character.creator}</span>
+                    <span style={conversionStyle}>
+                        {getPercent(character)}% conversion
+                    </span>
+                </>}
             </div>
         </div>
     })}